# docs/swagger.yaml
basePath: /
definitions:
  contracts.ErrPayload:
    properties:
      code:
        type: string
      description:
        type: string
    type: object
  contracts.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/contracts.ErrPayload'
    type: object
...
  exercises.CreateExerciseRequest:
    properties:
      description:
        type: string
      title:
        type: string
      trainer_id:
        type: string
      trainingPlanID:
        type: integer
    required:
    - description
    - title
    - trainer_id
    type: object
  exercises.DeleteExerciseRequest:
    properties:
      exerciseID:
        type: integer
      trainer_id:
        type: string
      trainingPlanID:
        type: integer
    required:
    - trainer_id
    type: object
  exercises.UpdateExerciseRequest:
    properties:
      description:
        type: string
      exerciseID:
        type: integer
      title:
        type: string
      trainer_id:
        type: string
      trainingPlanID:
        type: integer
    required:
    - trainer_id
    type: object
  goals.CreateGoalRequest:
    properties:
      deadline:
        type: string
      subtype:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
      value:
        type: integer
    required:
    - deadline
    - title
    - type
    - user_id
    - value
    type: object
  goals.DeleteGoalRequest:
    properties:
      goalID:
        type: integer
      user_id:
        type: string
    required:
    - user_id
    type: object
  goals.GetGoalsResponse:
    properties:
      goals:
        items:
          $ref: '#/definitions/models.Goal'
        type: array
    type: object
  goals.UpdateGoalRequest:
    properties:
      deadline:
        type: string
      goalID:
        type: integer
      title:
        type: string
      user_id:
        type: string
      value:
        type: integer
    required:
    - user_id
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Exercise:
    properties:
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      trainingPlanID:
        type: integer
      trainingPlanVersion:
        type: integer
    type: object
...
  models.Goal:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      id:
        type: integer
      progress:
        type: integer
      subtype:
        type: string
      title:
        type: string
      type:
        type: string
      user_id:
        type: string
      value:
        type: integer
    type: object
...
host: fiufit-trainings.fly.dev
info:
  contact: {}
  description: Fiufit's Trainings service documentation. This service manages training
    plans and its sessions, goals and reviews, etc.
  title: Fiufit Trainings API
  version: dev
paths:
  /{version}/goals:
    get:
      consumes:
      - application/json
      description: Get goals  with pagination filtered by their deadline, type, subtype
        for a given userID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: User ID
        in: query
        name: userID
        required: true
        type: string
      - description: Goal Type
        in: query
        name: type
        type: string
      - description: Goal Subtype
        in: query
        name: subtype
        type: string
      - description: Goal Deadline
        in: query
        name: deadline
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/goals.GetGoalsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Get goals by different query params with pagination
      tags:
      - goals
    post:
      consumes:
      - application/json
      description: Creates a new training goal for a given userID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/goals.CreateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Creates a new goal
      tags:
      - goals
  /{version}/goals/{goalID}:
    delete:
      consumes:
      - application/json
      description: Delete a goal for a given goalID, validating that the goal belongs
        to the userID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/goals.DeleteGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Deletes a goal
      tags:
      - goals
    get:
      consumes:
      - application/json
      description: Gets a goal for a given goalID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Goal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Gets a goal by ID
      tags:
      - goals
    patch:
      consumes:
      - application/json
      description: Updates a goal for a given goalID, validating that the goal was
        created by the user
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: Goal ID
        in: path
        name: goalID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/goals.UpdateGoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Updates a goal
      tags:
      - goals
...
  /{version}/trainings/{trainingID}/exercises:
    post:
      consumes:
      - application/json
      description: Creates a new exercise for a given trainingID, validating the training
        creator.
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: ID of the training to create the exercise for
        in: path
        name: trainingID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/exercises.CreateExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Exercise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Creates a new exercise
      tags:
      - exercises
  /{version}/trainings/{trainingID}/exercises/{exerciseID}:
    delete:
      consumes:
      - application/json
      description: Delete an exercise for a given trainingID and exerciseID, validating
        the training creator.
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: ID of the training to delete the exercise from
        in: path
        name: trainingID
        required: true
        type: integer
      - description: ID of the exercise to delete
        in: path
        name: exerciseID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/exercises.DeleteExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Delete an exercise
      tags:
      - exercises
    get:
      consumes:
      - application/json
      description: Get an exercise for a given trainingID and exerciseID
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: ID of the training to create the exercise for
        in: path
        name: trainingID
        required: true
        type: integer
      - description: ID of the exercise to delete
        in: path
        name: exerciseID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Exercise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Get exercise
      tags:
      - exercises
    patch:
      consumes:
      - application/json
      description: Update an exercise for a given trainingID and exerciseID, validating
        the training creator
      parameters:
      - description: API Version
        in: path
        name: version
        required: true
        type: string
      - description: ID of the training to update the exercise from
        in: path
        name: trainingID
        required: true
        type: integer
      - description: ID of the exercise to delete
        in: path
        name: exerciseID
        required: true
        type: integer
      - description: Body params
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/exercises.UpdateExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Important Note: OK responses are wrapped in {"data": ... }'
          schema:
            $ref: '#/definitions/models.Exercise'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contracts.ErrResponse'
      summary: Update exercise
      tags:
      - exercises
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"